Move some test helpers into a separate trait
