+++ phpunit.xml
@@ -1,4 +1,4 @@
-<phpunit backupGlobals="false" colors="true">
+<phpunit bootstrap="tests/Helpers.php" backupGlobals="false" colors="true">
   <testsuites>
     <testsuite>
       <directory suffix="_test.php">tests</directory>
